ingestion/requirements.txt
PyGithub==2.2.0
openai==0.27.4
langchain==0.0.278
redis==5.0.1
tiktoken==0.6.0
python-dotenv==1.0.1
chardet==5.2.0